Correct #37 (manage Repositories) and code clean
